feat: Add ability to skip to next line in custom board string using slash
