+++ src/game_modes/game_modifiers/start_board.rs
@@ -1,4 +1,4 @@
-use falling_tetromino_engine::{Game, GameAccess, GameBuilder, GameModifier};
+use falling_tetromino_engine::{Board, Game, GameAccess, GameBuilder, GameModifier};
 
 #[derive(
     PartialEq, Eq, PartialOrd, Ord, Hash, Clone, Debug, serde::Serialize, serde::Deserialize,
@@ -31,9 +31,56 @@ impl GameModifier for StartBoard {
     }
 
     fn on_game_built(&mut self, game: GameAccess) {
-        let start_board =
-            crate::application::NewGameSettings::decode_board(self.encoded_board.as_str());
+        let start_board = Self::decode_board(self.encoded_board.as_str());
 
         game.state.board = start_board;
+    }
+}
+
+impl StartBoard {
+    #[allow(dead_code)]
+    pub fn encode_board(board: &Board) -> String {
+        board
+            .iter()
+            .map(|line| {
+                line.iter()
+                    .map(|tile| if tile.is_some() { 'O' } else { ' ' })
+                    .collect::<String>()
+            })
+            .collect::<String>()
+            .trim_end()
+            .to_owned()
+    }
+
+    pub fn decode_board(board_str: &str) -> Board {
+        let grey_tile = Some(std::num::NonZeroU8::try_from(254).unwrap());
+
+        let mut new_board = Board::default();
+
+        let mut chars = board_str.chars();
+
+        'lines: for line in &mut new_board {
+            'tiles: for tile in line {
+                'chars: for char in chars.by_ref() {
+                    if char == '/' {
+                        // Slash = jump to next line (i.e. above).
+                        continue 'lines;
+                    } else if char == '\n' {
+                        // Newline = ignore, stay at tile but move on to next char.
+                        continue 'chars;
+                    } else if char == ' ' {
+                        // Space = empty tile.
+                        *tile = None;
+                        continue 'tiles;
+                    } else {
+                        // Otherwise = filled tile.
+                        *tile = grey_tile;
+                        continue 'tiles;
+                    }
+                }
+            }
+        }
+
+        new_board
     }
 }
